Support wake_usb hardware hack and update key order
This supports using D0 and D10 to briefly ground D+ and reset usb.

Solder D0 and D10 both to D+ in order to use this workaround.

TODO: Check if this can be done with one GPIO by changing current
settings from 20mA to 40mA (the usb pins are 40mA capable).

TODO: See if this can be done purely in software / firmware.